Improve agent heuristics (#19)
* Fix emitting position updates

A better approach to this would be to have a time step observable... but oh well.

* Combine heuristics to get an item

* Simplify checkout and exit heuristic

* Support simpler heuristic configuration

If there is only one heuristic in a set, it does not have to be wrapped in a list - this wrapping is done in the factory.

* Render number of items an agent has

* Add random item heuristic

* Update choosing best direction to be reactive

* Add commenting for re-worked time step

* Avoid duplicating name in heuristic dictionary

* Add wandering heuristic and agents

* Make navigation heuristic abstract

* Support spectator agents

Spectator agents are excluded from gameplay on the server-side. This allows an observer client to observe the game.

Decoupling graphics from a specific agent, for now. I would add this back when humans can control an agent.

* Remove hard-coded store configuration

Passed in as an argument when running the server

* Add additional logging

This was used to help diagnose and resolve issues when running many concurrent agents.

* Fixes for concurrency issues