Random agent configuration (#26)
* Additional heuristics for shoplifters

And the base for random configuration, though I don't feel this allows us to create "character" agents in the way I would like.

* More proximity heuristics

Allowing agents to pick up random nearby items.

If I were to do all this again, I feel like I would rearchitect it to allow heuristics to choose items based on "filters"... rather than repeating myself a bunch.

* Randomly create shopper configuration

An area to improve navigation: line-of-sight check for a destination, i.e. do I need a waypoint.

* Additional checks for valid shopper state

* Prevent agents from picking up more than one

* Fix for heuristic set scope

* Distance heuristics

It started to make sense for this to be abstract - didn't feel right to be repeating the corner logic.

* Fixes for distance heuristics

* Attempt to fix agents getting stuck

This seems to be caused by spontaneity heuristics completing an intermediary goal followed by returning to the agent's main goal.

Choosing to end the set of heuristics altogether once a spontaneity heuristic has been activated and satisfied.

* Shoplifter movement

A lot of changes around trying to navigate towards a position here too.

* Fix heuristic names in configuration

* Tidy logic in distance heuristic

This was a mess before. Now it can be summed up in:
- Never pick a shelf or door
- If you're in a corner, you're done

* Fixes for invalid agent states

If stuck in a corner, an agent could end up picking up another item or paying for an item.

* Add timeout back to spawner

This is a precaution for dealing with batches of agents at scale.

If there is an error and we don't disconnect, this should not result in the data generation stalling.

* Remove Pygame dependency from main

We should only be using the Pygame clock and 60 FPS when rendering graphics.

Now these are split out and Main is used for auto-generation, this significantly limits the speed of the program since network calls are only made 60 times per second (and all clients need to reach consensus).

Using sleep(0.001) as shown in PodSixNet example clients.